fix-errs2
